Radio jammer blocks ai interacion
