+++ Resources/Prototypes/Entities/Objects/Tools/jammer.yml
@@ -2,7 +2,7 @@
   name: radio jammer
   parent: [BaseItem, BaseSyndicateContraband]
   id: RadioJammer
-  description: This device will disrupt any nearby outgoing radio communication as well as suit sensors when activated.
+  description: This device will disrupt any nearby outgoing radio communication, suit sensors and the AIs ability to control electronics when activated.
   components:
   - type: Sprite
     sprite: Objects/Devices/jammer.rsi
